repo: respect the default None value of MAINTAINER_KEYS
Issue: LIN1024-742

The verification of the repo codes are done by invoking 'git tag -v'
which means the currently used rev is a tag and the tag is signed
by the GPG key pair's private part.

In WRLinux, we don't do this. We don't deiliver a public key and we
don't create a tag for repo and sign it off via the corresponding
private key. And there's no need to do so, as repo is part of the
WRLinux product and the interity of repo is ensured by our delivery
process and the git repo.

This means we don't change the default value of the MAINTAINER_KEYS.
And the check_repo_verify function should respect its default value.

This default None value should be respected even out of the WRLinux
context, because if you don't give people any public key, of course
they cannot verify your tag.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>